Add credentials to localStorage
