+++ atd-vze/src/containers/DefaultLayout/DefaultLayout.js
@@ -39,10 +39,7 @@ class DefaultLayout extends Component {
       <div className="app">
         <AppHeader fixed>
           <Suspense fallback={this.loading()}>
-            <DefaultHeader
-              auth={this.props.auth}
-              onLogout={e => this.signOut(e)}
-            />
+            <DefaultHeader onLogout={e => this.signOut(e)} />
           </Suspense>
         </AppHeader>
         <div className="app-body">
